+++ src/components/ui/label.jsx
@@ -5,7 +5,7 @@ function Label({ className, ...props }) {
   return (
     <LabelPrimitive.Root
       className={cn(
-        'text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70',
+        'text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70 dark:text-gray-200',
         className,
       )}
       {...props}